Changes evaluation logic to be model centered.
