+++ hrc_speech_prediction/evaluation.py
@@ -21,9 +21,11 @@ class Evaluation(object):
     Runs a number of different evaluations (per-user, ...).
     """
 
-    def __init__(self, model_generator, data_path, n_grams=(1, 1), tfidf=False):
+    def __init__(self, model_generator, data_path, n_grams=(1, 1), tfidf=False,
+                 model_variations=None):
         self.data = TrainData.load(os.path.join(data_path, "train.json"))
         self.model_gen = model_generator
+        self.model_variations = model_variations
         self.X_context, self.context_actions = \
             features.get_context_features(self.data)
         self.X_speech, _ = features.get_bow_features(
@@ -40,7 +42,7 @@ class Evaluation(object):
     def check_indices(self, train_idx, test_idx):
         assert (not set(train_idx).intersection(test_idx))
 
-    def evaluate_on_one_participant(self, data_type="context"):
+    def evaluate_on_one_participant(self):
         """
         Leaves one participant out of training and then tests on them. Does
         this for each participant
@@ -54,11 +56,11 @@ class Evaluation(object):
             train_idx = [i for p in participants
                          for i in list(self.data.data[p].ids)
                          if not p == tst]
-            results[tst] = self._evaluate_on(train_idx, test_idx, data_type)
+            results[tst] = self._evaluate_each_model_on(train_idx, test_idx)
         self._print_result_table(results, "Participants")
         return [t for t in results.values()]
 
-    def evaluate_on_one_trial(self, data_type="context"):
+    def evaluate_on_one_trial(self):
         """
         Excludes on trial from training (i.e. A, B, or C) and uses these
         excluded trials as tests
@@ -74,10 +76,10 @@ class Evaluation(object):
             ]
             train_idx = [i for p in TRAIN_PARTICIPANTS
                          for i in self.data.data[p].ids if i not in test_idx]
-            results[tst] = self._evaluate_on(train_idx, test_idx, data_type)
+            results[tst] = self._evaluate_each_model_on(train_idx, test_idx)
         self._print_result_table(results, "Instruction")
 
-    def cross_validation(self, data_type="context"):
+    def cross_validation(self):
         """
         10-fold cross validation
         """
@@ -90,10 +92,10 @@ class Evaluation(object):
             next_i = min(i + step_size, n_samples)
             test_idx = [j for j in range(i, next_i)]
             train_idx = [j for j in range(0, n_samples) if j not in test_idx]
-            results.append(self._evaluate_on(train_idx, test_idx, data_type))
-        self._print_global_results(results)
+            results.append(self._evaluate_each_model_on(train_idx, test_idx))
+        self._print_result_table(dict(enumerate(results)), '10-fold CV')
 
-    def new_participant(self, data_type="context"):
+    def new_participant(self):
         print("Testing on pilot participant...")
         results = {}
         trials = ["A", "D", "T"]
@@ -107,19 +109,16 @@ class Evaluation(object):
                 for trial in self.data.data[part] for i in trial.ids
                 if t == trial.instruction
             ]
-            results[t] = self._evaluate_on(train_idx, test_idx, data_type)
+            results[t] = self._evaluate_each_model_on(train_idx, test_idx)
         self._print_result_table(results, 'Instruction')
 
     def evaluate_all(self):
-        for data_type in ["context", "speech", "both"]:
-            print("---------------testing on {}---------------"
-                  .format(data_type))
-            self.evaluate_on_one_participant(data_type)
-            self.evaluate_on_one_trial(data_type)
-            self.cross_validation(data_type)
-            self.new_participant(data_type)
+        self.evaluate_on_one_participant()
+        self.evaluate_on_one_trial()
+        self.cross_validation()
+        self.new_participant()
 
-    def evaluate_incremental_learning(self, data_type):
+    def evaluate_incremental_learning(self):
         classes = np.unique((self.data.labels))
         score = []
         for i in range(0, self.X_context.shape[0] - 1):
@@ -131,7 +130,6 @@ class Evaluation(object):
             test_Y = self.get_labels([i + 1])
             model = self.model_gen(
                 self.context_actions,
-                features=data_type,
                 randomize_context=.25
             ).partial_fit(train_X[0], train_X[1], train_Y, classes)
             prediction = model.predict(*test_X)
@@ -148,7 +146,19 @@ class Evaluation(object):
         plt.plot(xnew, score_smooth)
         plt.show()
 
-    def _evaluate_on(self, train_idx, test_idx, data_type):
+    @property
+    def _variations(self):
+        if self.model_variations is None:
+            return {'Accuracy': {}}
+        else:
+            return self.model_variations
+
+    def _evaluate_each_model_on(self, train_idx, test_idx):
+        return {m: self._evaluate_on(train_idx, test_idx,
+                                     model_params=self._variations[m])
+                for m in self._variations}
+
+    def _evaluate_on(self, train_idx, test_idx, model_params):
         self.check_indices(train_idx, test_idx)
         # Train set
         train_Xs = self.get_Xs(train_idx)
@@ -158,22 +168,35 @@ class Evaluation(object):
         test_Y = self.get_labels(test_idx)
         # Train
         model = self.model_gen(self.context_actions,
-                               features=data_type,
-                               randomize_context=.25
+                               randomize_context=.25,
+                               **model_params
                                ).fit(train_Xs[0], train_Xs[1], train_Y)
         # Evaluate
         prediction = model.predict(*test_Xs)
         return metrics.accuracy_score(
             test_Y, prediction, normalize=True, sample_weight=None)
 
-    def _print_result_table(self, results, key_title):
-        w = max(len(key_title), len("Accuracy"))
-        print("{:<{w}}: {} ".format(key_title, " ".join(
-            ["{:^7}".format(k) for k in results.keys()]), w=w))
-        print("{:<{w}}: {} ".format("Accuracy", " ".join(
-            ["{:^7.2f}".format(v) for v in results.values()]), w=w))
-        self._print_global_results(results.values())
+    def _reorder_by_model(self, results):
+        return {m: {k: results[k][m] for k in results} for m in self._variations}
 
-    def _print_global_results(self, results):
-        print("Average {:.3f}, std: {:.4f}\n".format(np.average(results),
-                                                     np.std(results)))
+    def _print_result_table(self, results, key_title):
+        results = self._reorder_by_model(results)
+        result_keys = list(results.values())[0].keys()
+        txt_cell = "{:^7}"
+        num_cell = "{:^7.2f}"
+        avg_cell = "{:^7.3f}"
+        std_cell = "{:^7.4f}"
+        w = max(len(key_title), len("Accuracy"))
+        print("{:<{w}}: {} ".format(
+            key_title,
+            " ".join([txt_cell.format(k) for k in result_keys] +
+                     [txt_cell.format("AVG."), txt_cell.format("STD.")]),
+            w=w))
+        for m in self._variations:
+            print("{:<{w}}: {} ".format(
+                m,
+                " ".join(
+                    [num_cell.format(v) for v in results[m].values()] +
+                    [avg_cell.format(np.average(results[m].values())),
+                     std_cell.format(np.std(results[m].values()))]),
+                w=w))
